tests: relocate all notification plugin tests
